+++ docs/changelog.rst
@@ -7,6 +7,7 @@ v0.8.3
 * Added a webapp manifest
 * Added a Heroku demo deployment
 * Improved some API queries
+* Fixed chapter view tracking
 
 v0.8.2
 ^^^^^^